pipecat-test/requirements.txt
# VoxKit Pipecat Test — Python dependencies
# Install: pip install -r requirements.txt

# Core framework with required extras
pipecat-ai[daily,openai,soniox,silero]>=0.0.98

# Web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP client (for Daily API calls)
httpx>=0.27.0

# Twilio (for call forwarding from bot.py)
twilio>=9.0.0

# Environment
python-dotenv>=1.0.0
